+++ src/main/java/com/pokemon/repository/DresseurRepository.java
@@ -3,5 +3,5 @@ package com.pokemon.repository;
 import com.pokemon.entity.Dresseurs;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface DresseurRepository extends JpaRepository<Dresseurs, Long> {
+public interface DresseurRepository extends JpaRepository<Dresseurs, Integer> {
 }
