+++ src/main/java/com/pokemon/repository/NatureRepository.java
@@ -3,5 +3,5 @@ package com.pokemon.repository;
 import com.pokemon.entity.Natures;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface NatureRepository extends JpaRepository<Natures, Long> {
+public interface NatureRepository extends JpaRepository<Natures, Integer> {
 }
