+++ src/main/java/com/pokemon/repository/PokeballRepository.java
@@ -3,5 +3,5 @@ package com.pokemon.repository;
 import com.pokemon.entity.Pokeballs;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PokeballRepository extends JpaRepository<Pokeballs, Long> {
+public interface PokeballRepository extends JpaRepository<Pokeballs, Integer> {
 }
