+++ src/main/java/com/pokemon/repository/PokemonShinyRepository.java
@@ -3,5 +3,5 @@ package com.pokemon.repository;
 import com.pokemon.entity.PokemonShiny;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PokemonShinyRepository extends JpaRepository<PokemonShiny, Long> {
+public interface PokemonShinyRepository extends JpaRepository<PokemonShiny, Integer> {
 }
