+++ src/main/java/com/pokemon/repository/TypeRepository.java
@@ -3,5 +3,5 @@ package com.pokemon.repository;
 import com.pokemon.entity.Types;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TypeRepository extends JpaRepository<Types, Long> {
+public interface TypeRepository extends JpaRepository<Types, Integer> {
 }
